+++ drivers/misc/mediatek/eccci/hif/ccci_dpmaif_drv_v3.c
@@ -1178,7 +1178,7 @@ static void drv3_dump_register(int buf_type)
 	len = DPMAIF_PD_UL_ADD_DESC_CH - DPMAIF_PD_UL_ADD_DESC + 4;
 	CCCI_BUF_LOG_TAG(0, buf_type, TAG,
 		"dump AP DPMAIF Tx pdn; pd_ul_base register -> (start addr: 0x%llX, len: %d):\n",
-		dpmaif_ctl->pd_ul_base + NRL2_DPMAIF_UL_ADD_DESC, len);
+		(unsigned long long)dpmaif_ctl->pd_ul_base + NRL2_DPMAIF_UL_ADD_DESC, len);
 	ccci_util_mem_dump(buf_type,
 		dpmaif_ctl->pd_ul_base + NRL2_DPMAIF_UL_ADD_DESC, len);
 
@@ -1186,14 +1186,15 @@ static void drv3_dump_register(int buf_type)
 		len = DPMAIF_AO_UL_CHNL3_STA_6985 - DPMAIF_AO_UL_CHNL0_STA_6985 + 4;
 		CCCI_BUF_LOG_TAG(0, buf_type, TAG,
 			"dump AP DPMAIF Tx ao; ao_ul_base register -> (start addr: 0x%llX, len: %d):\n",
-			dpmaif_ctl->ao_ul_base + DPMAIF_AO_UL_CHNL0_STA_6985, len);
+			(unsigned long long)dpmaif_ctl->ao_ul_base + DPMAIF_AO_UL_CHNL0_STA_6985,
+				len);
 		ccci_util_mem_dump(buf_type,
 			dpmaif_ctl->ao_ul_base + DPMAIF_AO_UL_CHNL0_STA_6985, len);
 	} else {
 		len = DPMAIF_AO_UL_CHNL3_STA - DPMAIF_AO_UL_CHNL0_STA + 4;
 		CCCI_BUF_LOG_TAG(0, buf_type, TAG,
 			"dump AP DPMAIF Tx ao; ao_ul_base register -> (start addr: 0x%llX, len: %d):\n",
-			dpmaif_ctl->ao_ul_base + DPMAIF_AO_UL_CHNL0_STA, len);
+			(unsigned long long)dpmaif_ctl->ao_ul_base + DPMAIF_AO_UL_CHNL0_STA, len);
 		ccci_util_mem_dump(buf_type,
 			dpmaif_ctl->ao_ul_base + DPMAIF_AO_UL_CHNL0_STA, len);
 	}
@@ -1201,40 +1202,40 @@ static void drv3_dump_register(int buf_type)
 	len = DPMAIF_PD_DL_MISC_CON0 - DPMAIF_PD_DL_BAT_INIT + 4;
 	CCCI_BUF_LOG_TAG(0, buf_type, TAG,
 		"dump AP DPMAIF Rx pdn; pd_dl_base register -> (start addr: 0x%llX, len: %d):\n",
-		dpmaif_ctl->pd_dl_base + DPMAIF_PD_DL_BAT_INIT, len);
+		(unsigned long long)dpmaif_ctl->pd_dl_base + DPMAIF_PD_DL_BAT_INIT, len);
 	ccci_util_mem_dump(buf_type,
 		dpmaif_ctl->pd_dl_base + DPMAIF_PD_DL_BAT_INIT, len);
 
 	len = DPMAIF_PD_DL_DBG_STA14 - DPMAIF_PD_DL_STA0 + 4;
 	CCCI_BUF_LOG_TAG(0, buf_type, TAG,
 		"dump AP DPMAIF Rx pdn; pd_dl_base register -> (start addr: 0x%llX, len: %d):\n",
-		dpmaif_ctl->pd_dl_base + DPMAIF_PD_DL_STA0, len);
+		(unsigned long long)dpmaif_ctl->pd_dl_base + DPMAIF_PD_DL_STA0, len);
 	ccci_util_mem_dump(buf_type,
 		dpmaif_ctl->pd_dl_base + DPMAIF_PD_DL_STA0, len);
 
 	CCCI_BUF_LOG_TAG(0, buf_type, TAG,
 		"dump AP DPMAIF dma_rd; pd_dl_base register -> (start addr: 0x%llX, len: %d):\n",
-		dpmaif_ctl->pd_dl_base + 0x100, 0xC8);
+		(unsigned long long)dpmaif_ctl->pd_dl_base + 0x100, 0xC8);
 	ccci_util_mem_dump(buf_type,
 		dpmaif_ctl->pd_dl_base + 0x100, 0xC8);
 
 	CCCI_BUF_LOG_TAG(0, buf_type, TAG,
 		"dump AP DPMAIF dma_wr; pd_dl_base register -> (start addr: 0x%llX, len: %d):\n",
-		dpmaif_ctl->pd_dl_base + 0x200, 0x58 + 4);
+		(unsigned long long)dpmaif_ctl->pd_dl_base + 0x200, 0x58 + 4);
 	ccci_util_mem_dump(buf_type,
 		dpmaif_ctl->pd_dl_base + 0x200, 0x58 + 4);
 
 	len = DPMAIF_AO_DL_FRGBAT_STA2 - DPMAIF_AO_DL_PKTINFO_CONO + 4;
 	CCCI_BUF_LOG_TAG(0, buf_type, TAG,
 		"dump AP DPMAIF Rx ao; ao_dl_base register -> (start addr: 0x%llX, len: %d):\n",
-		dpmaif_ctl->ao_dl_base + DPMAIF_AO_DL_PKTINFO_CONO, len);
+		(unsigned long long)dpmaif_ctl->ao_dl_base + DPMAIF_AO_DL_PKTINFO_CONO, len);
 	ccci_util_mem_dump(buf_type,
 		dpmaif_ctl->ao_dl_base + DPMAIF_AO_DL_PKTINFO_CONO, len);
 
 	len = DPMAIF_PD_AP_CODA_VER - DPMAIF_PD_AP_UL_L2TISAR0 + 4;
 	CCCI_BUF_LOG_TAG(0, buf_type, TAG,
 		"dump AP DPMAIF MISC pdn; pd_misc_base register -> (start addr: 0x%llX, len: %d):\n",
-		dpmaif_ctl->pd_misc_base + DPMAIF_PD_AP_UL_L2TISAR0, len);
+		(unsigned long long)dpmaif_ctl->pd_misc_base + DPMAIF_PD_AP_UL_L2TISAR0, len);
 	ccci_util_mem_dump(buf_type,
 		dpmaif_ctl->pd_misc_base + DPMAIF_PD_AP_UL_L2TISAR0, len);
 
@@ -1242,7 +1243,7 @@ static void drv3_dump_register(int buf_type)
 	DPMA_WRITE_PD_MISC(DPMAIF_PD_AP_CG_EN, 0x36);
 	CCCI_BUF_LOG_TAG(0, buf_type, TAG,
 		"dump AP DPMAIF SRAM pdn; pd_sram_base register -> (start addr: 0x%llX, len: %d):\n",
-		dpmaif_ctl->pd_sram_base + 0x00, 0x184);
+		(unsigned long long)dpmaif_ctl->pd_sram_base + 0x00, 0x184);
 	ccci_util_mem_dump(buf_type,
 		dpmaif_ctl->pd_sram_base + 0x00, 0x184);
